Add option to Makefile for a lot of warnings, cache BSP JSON parse re…

+++ src/bsp/CScaledBSP.cpp
@@ -16,7 +16,7 @@
 void CScaledBSP::IScaledBSP(Fixed scale, short resId, CAbstractActor *anActor, short aPartCode) {
     //OSErr iErr;
     Vector *p;
-    short i;
+    size_t i;
 
     CSmartPart::ISmartPart(resId, anActor, aPartCode);
